refactor, write activation
